updated readme

+++ src/HTTPProxyServer.py
@@ -40,12 +40,10 @@ class MITMProxyServer:
 
 
 def prepare_log_dir(log):
-        if log is not None:
-            if not os.path.isabs(log):
-                log = os.path.join('../', log)
-            if os.path.exists(log):
-                shutil.rmtree(log)
-            os.mkdir(log)
+    if log is not None and log != '.':
+        if os.path.exists(log):
+            shutil.rmtree(log)
+        os.mkdir(log)
 
 
 class Worker:
